requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyserial>=3.5
pandas==1.3.5
python-multipart>=0.0.6
